pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "npm-cli"
version = "1.0.0"
description = "Full-featured CLI for Nginx Proxy Manager - manage proxy hosts, certificates, streams, and more"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Oriol Rius", email = "oriol@joor.net" }
]
keywords = [
    "nginx",
    "proxy",
    "manager",
    "cli",
    "ssl",
    "certificates",
    "letsencrypt",
    "reverse-proxy"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy Servers",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities"
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "httpx>=0.24",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "keyring>=24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
npm-cli = "npm_cli.main:cli"

[project.urls]
Homepage = "https://github.com/oriolrius/npm-cli"
Repository = "https://github.com/oriolrius/npm-cli"
Issues = "https://github.com/oriolrius/npm-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/npm_cli"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.9"
strict = true
